# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "目前仅仅是一个demo",
        "title": "Rest API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/GetAllButtonByMenuIdForTransfer": {
            "post": {
                "description": "根据菜单ID获取按钮清单 请求主体: Null  成功输出[]TransferModel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Button"
                ],
                "summary": "根据菜单ID获取按钮清单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]TransferModel},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetAllButtonForTransfer": {
            "post": {
                "description": "获取所有菜单清单 请求主体: Null  成功输出[]TransferModel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Button"
                ],
                "summary": "获取所有菜单清单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]TransferModel},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetAllDeptForTransfer": {
            "post": {
                "description": "获取所有部门清单 请求主体: Null  成功输出[]TransferModel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept"
                ],
                "summary": "获取所有部门清单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]TransferModel},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetButtonByMenuIdAndUserId": {
            "post": {
                "description": "根据用户ID获取用户菜单信息 请求主体:  base64(userId=aaaa\u0026menuId=aaaa)  成功输出[]Button",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Button"
                ],
                "summary": "根据用户ID和菜单ID获取相应权限的按钮信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "menuId",
                        "name": "menuId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]Button},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetButtonByMenuIdRoleIdForTransfer": {
            "post": {
                "description": "根据菜单ID和权限ID获取按钮信息 请求主体: base64(menuId=aaaa\u0026roleId=bbbb) 成功输出字符串",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Button"
                ],
                "summary": "根据菜单ID和权限ID获取按钮信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "menuId",
                        "name": "menuId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "roleId",
                        "name": "roleId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":\"aaa,bbb,ccc\",\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetButtonByMenuIdnForTransfer": {
            "post": {
                "description": "根据菜单ID获取按钮信息 请求主体: base64(menuId=aaaa) 成功输出字符串",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Button"
                ],
                "summary": "根据菜单ID获取按钮信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "menuId",
                        "name": "menuId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":\"aaa,bbb,ccc\",\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetDeptByUserIdForTransfer": {
            "post": {
                "description": "根据用户ID获取用户部门信息 请求主体: base64(userId=aaaa) 成功输出字符串",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept"
                ],
                "summary": "根据用户ID获取用户部门信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":\"aaa,bbb,ccc\",\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}